# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /businesses:
    get:
      operationId: listBusinesses
      tags: [businesses]
      summary: List all businesses
      responses:
        "200":
          description: List of businesses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Business"
    post:
      operationId: createBusiness
      tags: [businesses]
      summary: Create or update a business profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBusinessBody"
      responses:
        "201":
          description: Created business
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Business"
  /businesses/{id}:
    get:
      operationId: getBusiness
      tags: [businesses]
      summary: Get a business by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Business with latest audit summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BusinessWithSummary"
        "404":
          description: Not found
    put:
      operationId: updateBusiness
      tags: [businesses]
      summary: Update a business profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBusinessBody"
      responses:
        "200":
          description: Updated business
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Business"
  /businesses/{id}/audits:
    get:
      operationId: listAudits
      tags: [audits]
      summary: List all audits for a business
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of audits
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Audit"
    post:
      operationId: createAudit
      tags: [audits]
      summary: Submit a carbon audit for a business
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAuditBody"
      responses:
        "201":
          description: Created audit with calculated footprint
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditWithBreakdown"
  /businesses/{id}/audits/latest:
    get:
      operationId: getLatestAudit
      tags: [audits]
      summary: Get the latest audit result with full breakdown
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Latest audit with breakdown
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditWithBreakdown"
        "404":
          description: No audit found
  /businesses/{id}/action-plan:
    get:
      operationId: getActionPlan
      tags: [action-plans]
      summary: Get AI-generated action plan for a business
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Action plan with prioritized steps
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionPlan"
        "404":
          description: No action plan yet
    post:
      operationId: generateActionPlan
      tags: [action-plans]
      summary: Generate a new AI action plan based on latest audit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Generated action plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionPlan"
  /businesses/{id}/action-plan/items/{itemId}:
    patch:
      operationId: updateActionItem
      tags: [action-plans]
      summary: Update the status of an action item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateActionItemBody"
      responses:
        "200":
          description: Updated action item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionItem"
...
components:
  schemas:
...
    Business:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        sector:
          type: string
          description: "Business sector (e.g. kuliner, fashion, retail, manufaktur, jasa)"
        location:
          type: string
        employeeCount:
          type: integer
        description:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - sector
        - location
        - employeeCount
        - createdAt
        - updatedAt
    BusinessWithSummary:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        sector:
          type: string
        location:
          type: string
        employeeCount:
          type: integer
        description:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        latestAuditId:
          type: integer
          nullable: true
        latestTotalEmissions:
          type: number
          nullable: true
          description: "Total CO2e in tonnes per year"
        emissionsTrend:
          type: number
          nullable: true
          description: "Percentage change vs previous audit"
        actionItemsCompleted:
          type: integer
        actionItemsTotal:
          type: integer
      required:
        - id
        - name
        - sector
        - location
        - employeeCount
        - createdAt
        - updatedAt
        - actionItemsCompleted
        - actionItemsTotal
    CreateBusinessBody:
      type: object
      properties:
        name:
          type: string
        sector:
          type: string
        location:
          type: string
        employeeCount:
          type: integer
        description:
          type: string
      required:
        - name
        - sector
        - location
        - employeeCount
    Audit:
      type: object
      properties:
        id:
          type: integer
        businessId:
          type: integer
        totalEmissions:
          type: number
          description: "Total CO2e in tonnes per year"
        period:
          type: string
          description: "e.g. 2024-Q1, 2024-Annual"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - businessId
        - totalEmissions
        - period
        - createdAt
    AuditWithBreakdown:
      type: object
      properties:
        id:
          type: integer
        businessId:
          type: integer
        totalEmissions:
          type: number
        period:
          type: string
        createdAt:
          type: string
          format: date-time
        energyEmissions:
          type: number
          description: "CO2e from electricity/energy"
        transportEmissions:
          type: number
          description: "CO2e from transportation"
        wasteEmissions:
          type: number
          description: "CO2e from waste"
        supplyChainEmissions:
          type: number
          description: "CO2e from supply chain"
        electricityKwh:
          type: number
        fuelLiters:
          type: number
        wasteKg:
          type: number
        aiInsights:
          type: string
          description: "AI-generated insights about the audit"
      required:
        - id
        - businessId
        - totalEmissions
        - period
        - createdAt
        - energyEmissions
        - transportEmissions
        - wasteEmissions
        - supplyChainEmissions
        - electricityKwh
        - fuelLiters
        - wasteKg
    CreateAuditBody:
      type: object
      properties:
        period:
          type: string
          description: "Reporting period e.g. 2024-Annual"
        electricityKwh:
          type: number
          description: "Monthly electricity consumption in kWh"
        fuelLiters:
          type: number
          description: "Monthly fuel consumption in liters"
        wasteKg:
          type: number
          description: "Monthly waste generated in kg"
        supplyChainSpendIdr:
          type: number
          description: "Monthly supply chain spend in IDR"
        vehicleCount:
          type: integer
          description: "Number of operational vehicles"
        deliveriesPerMonth:
          type: integer
          description: "Number of deliveries per month"
      required:
        - period
        - electricityKwh
        - fuelLiters
        - wasteKg
    ActionPlan:
      type: object
      properties:
        id:
          type: integer
        businessId:
          type: integer
        auditId:
          type: integer
        summary:
          type: string
          description: "AI-generated summary of the plan"
        totalPotentialReduction:
          type: number
          description: "Estimated total CO2e reduction in tonnes"
        createdAt:
          type: string
          format: date-time
        items:
          type: array
          items:
            $ref: "#/components/schemas/ActionItem"
      required:
        - id
        - businessId
        - auditId
        - summary
        - totalPotentialReduction
        - createdAt
        - items
    ActionItem:
      type: object
      properties:
        id:
          type: integer
        planId:
          type: integer
        title:
          type: string
        description:
          type: string
        category:
          type: string
          description: "energi | transportasi | sampah | supply_chain"
        priority:
          type: string
          description: "quick_win | medium_term | long_term"
        estimatedReduction:
          type: number
          description: "Estimated CO2e reduction in tonnes"
        estimatedCostIdr:
          type: number
          description: "Estimated implementation cost in IDR"
        reasoning:
          type: string
          description: "AI explanation for why this action was recommended"
        status:
          type: string
          description: "pending | in_progress | completed"
        completedAt:
          type: string
          format: date-time
          nullable: true
      required:
        - id
        - planId
        - title
        - description
        - category
        - priority
        - estimatedReduction
        - reasoning
        - status
    UpdateActionItemBody:
      type: object
      properties:
        status:
          type: string
          description: "pending | in_progress | completed"
      required:
        - status
...