# lib/api-zod/src/generated/types/actionPlan.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ActionItem } from "./actionItem";

export interface ActionPlan {
  id: number;
  businessId: number;
  auditId: number;
  /** AI-generated summary of the plan */
  summary: string;
  /** Estimated total CO2e reduction in tonnes */
  totalPotentialReduction: number;
  createdAt: Date;
  items: ActionItem[];
}