# lib/api-zod/src/generated/types/audit.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Audit {
  id: number;
  businessId: number;
  /** Total CO2e in tonnes per year */
  totalEmissions: number;
  /** e.g. 2024-Q1, 2024-Annual */
  period: string;
  createdAt: Date;
}