# lib/api-zod/src/generated/types/business.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Business {
  id: number;
  name: string;
  /** Business sector (e.g. kuliner, fashion, retail, manufaktur, jasa) */
  sector: string;
  location: string;
  employeeCount: number;
  description?: string;
  createdAt: Date;
  updatedAt: Date;
}