# lib/api-zod/src/generated/types/businessWithSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface BusinessWithSummary {
  id: number;
  name: string;
  sector: string;
  location: string;
  employeeCount: number;
  description?: string;
  createdAt: Date;
  updatedAt: Date;
  latestAuditId?: number | null;
  /** Total CO2e in tonnes per year */
  latestTotalEmissions?: number | null;
  /** Percentage change vs previous audit */
  emissionsTrend?: number | null;
  actionItemsCompleted: number;
  actionItemsTotal: number;
}